test/fixtures/wpt/wasm/webapi/esm-integration/wasm-to-wasm-link-error.tentative.html

<!doctype html>
<title>Errors for linking WebAssembly module scripts</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    const test_load = async_test(
        "Link errors for imports between WebAssembly modules should be reported.");

    window.log = [];
    window.addEventListener("error", ev => {
      test_load.step(() => assert_equals(ev.error.constructor, WebAssembly.LinkError));
      log.push(ev.message);
    });

    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(log.length, 2);
      assert_equals(log[1], 1);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script type="module" src="./resources/wasm-import-error-from-wasm.wasm"
    onerror="unreachable()" onload="log.push(1)"></script>
