third_party/blink/web_tests/external/wpt/wasm/webapi/esm-integration/wasm-to-wasm-link-error.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Link errors for imports between WebAssembly modules should be reported.
  assert_equals: expected 2 but got 1
Harness: the test ran to completion.
